requirements.txt
torch==1.12.1
torchlibrosa==0.0.9
panns-inference==0.0.7
